refactor: @shared/components 경로를 @shared/ui로 변경 및 관련 파일 수정
