+++ src/features/coupon/ui/CouponList/index.tsx
@@ -1,5 +1,5 @@
 import * as S from './CouponList.styled';
-import Description from '@shared/components/Description';
+import Description from '@/shared/ui/Description';
 import { ClientCouponType } from '@entities/coupon/type/coupon.type';
 import CouponItem from '@features/coupon/ui/CouponItem';
 import { COUPON_RULE } from '@/entities/coupon/constants/couponRule';
