+++ src/main.tsx
@@ -3,7 +3,7 @@ import { BrowserRouter } from 'react-router-dom';
 import ReactDOM from 'react-dom/client';
 import App from '@app/App.tsx';
 import './styles/reset.css';
-import MobileLayout from '@shared/components/MobileLayout/MobileLayout.tsx';
+import MobileLayout from '@/shared/ui/MobileLayout/MobileLayout';
 
 ReactDOM.createRoot(document.getElementById('root')!).render(
   <React.StrictMode>
