+++ src/pages/order/OrderPage.tsx
@@ -1,6 +1,6 @@
-import BackButton from '@shared/components/BackButton/BackButton';
+import BackButton from '@/shared/ui/BackButton/BackButton';
 import { usePageNavigation } from '@app/hooks/usePageNavigation';
-import Header from '@shared/components/Header/Header';
+import Header from '@/shared/ui/Header/Header';
 import * as S from './OrderPage.styled';
 import OrderContent from '@widget/order/OrderContent/OrderContent';
