+++ src/shared/ui/BottomConfirmButton/BottomConfirmButton.styled.ts
