+++ src/shared/ui/CheckBox/index.tsx
