+++ src/shared/ui/Description/Description.styled.ts
