+++ src/shared/ui/Description/index.tsx
