+++ src/shared/ui/ErrorContainer/ErrorContainer.styled.ts
