+++ src/shared/ui/ErrorContainer/ErrorContainer.tsx
