+++ src/shared/ui/MobileLayout/MobileLayout.styled.ts
