+++ src/shared/ui/MobileLayout/MobileLayout.tsx
