+++ src/shared/ui/Modal/Modal.styled.ts
