+++ src/shared/ui/PriceInfoContainer/PriceBox/PriceBox.styled.ts
