+++ src/shared/ui/PriceInfoContainer/PriceBox/PriceTextBox/PriceTextBox.styled.ts
