+++ src/shared/ui/PriceInfoContainer/PriceBox/PriceTextBox/PriceTextBox.tsx
