+++ src/shared/ui/PriceInfoContainer/PriceBox/index.tsx
