+++ src/shared/ui/PriceInfoContainer/PriceInfoContainer.styled.ts
