+++ src/shared/ui/PriceInfoContainer/index.tsx
